requirements.txt
fastapi>=0.110.0
uvicorn>=0.28.0
sqlalchemy>=2.0.28
alembic>=1.13.1
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4.0.0
python-multipart>=0.0.9
pydantic>=2.6.4
pydantic-settings>=2.2.1
stripe>=8.6.0
httpx>=0.27.0
